+++ backend/TimeTile/TimeTile.API/Users/Requests/ICreateUserRequest.cs
@@ -1,0 +1,11 @@
+﻿namespace TimeTile.API.Users.Requests
+{
+    public interface ICreateUserRequest
+    {
+        string Firstname { get; }
+        string Lastname { get; }
+        string PhoneNumber { get; }
+        string HomeAddress { get; }
+        DateOnly BirthDate { get; }
+    }
+}
